Convert temperature to double

+++ src/oscillator.c
@@ -43,7 +43,7 @@ int oscillator_save(struct oscillator *oscillator)
 	return oscillator->class->save(oscillator);
 }
 
-int oscillator_get_temp(struct oscillator *oscillator, uint16_t *temp)
+int oscillator_get_temp(struct oscillator *oscillator, double *temp)
 {
 	if (oscillator == NULL || temp == NULL)
 		return -EINVAL;
