+++ src/oscillator.h
@@ -19,7 +19,7 @@ typedef int (*oscillator_get_ctrl_cb)(struct oscillator *oscillator,
 		struct oscillator_ctrl *ctrl);
 typedef int (*oscillator_save_cb)(struct oscillator *oscillator);
 typedef int (*oscillator_get_temp_cb)(struct oscillator *oscillator,
-		uint16_t *temp);
+		double *temp);
 typedef int (*oscillator_apply_output_cb)(struct oscillator *oscillator,
 		struct od_output *output);
 typedef void (*oscillator_destroy_cb)(struct oscillator **oscillator);
@@ -62,7 +62,7 @@ int oscillator_set_dac_min(struct oscillator *oscillator, uint32_t dac_min);
 int oscillator_set_dac_max(struct oscillator *oscillator, uint32_t dac_max);
 int oscillator_get_ctrl(struct oscillator *oscillator, struct oscillator_ctrl *ctrl);
 int oscillator_save(struct oscillator *oscillator);
-int oscillator_get_temp(struct oscillator *oscillator, uint16_t *temp);
+int oscillator_get_temp(struct oscillator *oscillator, double *temp);
 int oscillator_apply_output(struct oscillator *oscillator, struct od_output *output);
 struct calibration_results * oscillator_calibrate(
 	struct oscillator *oscillator,
